+++ tools/daemon/command.do.pl
@@ -11,7 +11,7 @@
 # @(-) --timeout=<timeout>     command timeout in sec. [${timeout}]
 #
 # @(@) Note 1: a command is a simple string. The daemon is expected to (at least) acknowledge it.
-# @(@) Note 2: "daemon.pl" command and all its verbs only work on the local node.
+# @(@) Note 2: "daemon.pl" command and all its verbs only act on the local node.
 #
 # TheToolsProject - Tools System and Working Paradigm for IT Production
 # Copyright (©) 1998-2023 Pierre Wieser (see AUTHORS)