+++ tools/daemon/status.do.pl
@@ -12,7 +12,7 @@
 # @(-) --label=<name=value>    label(s) to be added to the published metric, may be specified several times or as comma-separated strings [${label}]
 #
 # @(@) Note 1: published telemetries exhibit the very same metrics than those periodically advertised by the daemon itself.
-# @(@) Note 2: "daemon.pl" command and all its verbs only work on the local node.
+# @(@) Note 2: "daemon.pl" command and all its verbs only act on the local node.
 #
 # TheToolsProject - Tools System and Working Paradigm for IT Production
 # Copyright (©) 1998-2023 Pierre Wieser (see AUTHORS)