+++ tools/daemon/stop.do.pl
@@ -11,7 +11,7 @@
 # @(-) --[no]wait              wait for actual termination [${wait}]
 # @(-) --timeout=<timeout>     timeout in sec. when waiting for termination [${timeout}]
 #
-# @(@) Note 1: "daemon.pl" command and all its verbs only work on the local node.
+# @(@) Note 1: "daemon.pl" command and all its verbs only act on the local node.
 #
 # TheToolsProject - Tools System and Working Paradigm for IT Production
 # Copyright (©) 1998-2023 Pierre Wieser (see AUTHORS)